Script changes and Narrative finishing
Made script changes to trigger zones script and how I was handling that to make it simpler. Also made changes to GameSystemBehavior with the NarrativeEvent and narrative state stuff to make things a little bit smoother in other scripts. Somewhat finished narrative with just reuniting Ash with Charizard. If we want to do a boss fight and portal will have to incorporate Morgan's fight in the end and add more narrative.